requirements.txt
altair
numpy
pandas
pydeck
streamlit
streamlit_pandas_profiling